ISAICP-3443: Remove stale collection classes.
